python-scheduler/requirements.txt
pandas
duckdb
schedule
python-dotenv
boto3
pyarrow
pyyaml
